+++ soslab_api/internal/Sensor/src/MLU.cpp
@@ -23,6 +23,7 @@ soslab::MLU::MLU()
 	{
 		frameData.echoNum = 1;
 		frameData.lidarId = 0;
+		frameData.timestamp.resize(32);
 		frameData.ambient.resize(numTotalSize);
 		frameData.depth[0].resize(numTotalSize);
 		frameData.intensity[0].resize(numTotalSize);
@@ -160,6 +161,7 @@ bool soslab::MLU::buildStreamData(const std::vector<uint8_t>& packetData)
 
 	FrameData& frameData = frameDataVec[numbering];
 	frameData.lidarId = numbering;
+	frameData.timestamp[vroll] = headerInfo.timestamp;
 
 	for (int r = 0; r < numSegmentRow; r++)
 	{