+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.0.1] - 2026-04-29
+
+### Added
+- Add timestamp information in FrameData structure
+- Add how to access timestamp in ROS1 / ROS2 examples
+
 ## [1.0.0] - 2026-03-31
 
 ### Added